apps/medical_records/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=41.0.8
bcrypt>=4.0.0
email-validator==2.1.0
PyJWT[crypto]==2.8.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
redis==5.0.1

# Google Cloud Services
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-cloud-aiplatform==1.38.1

# File Processing and OCR
Pillow==10.1.0
python-magic==0.4.27
PyMuPDF==1.26.3
pdf2image==1.16.3
pytesseract==0.3.13

# Data Processing
pandas==2.1.3
numpy==1.25.2

# JSON and Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and Monitoring
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
prometheus_client==0.19.0

# Medical Imaging and DICOM Processing
pydicom==2.4.3
pynetdicom==2.0.2
opencv-python==4.8.1.78
scikit-image==0.21.0

# AI/ML Dependencies for Medical Analysis
transformers>=4.36.0
torch>=2.1.1
scikit-learn>=1.3.0

# Message Queue and Event Streaming
confluent-kafka>=2.3.0
kafka-python>=2.0.2

# Resilience and Circuit Breaker
tenacity>=8.2.3
pybreaker>=1.0.1

# Health checks
healthcheck>=1.3.3

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
